+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog 
 ### 1.0.1 (10/01/16)
-* Less code used to removt HTTP and HTTPS protocols
+* Improve regex to find all protocols
+* Less code used to remove HTTP and HTTPS protocols
 * Comments added for code explanation
 
 ### 1.0.0 (09/05/16)